commit 25

+++ src/main/java/hello/board/service/CommentService.java
@@ -1,10 +1,12 @@
 package hello.board.service;
 
 import hello.board.domain.comment.Comment;
+import hello.board.repository.BoardRepository;
 import hello.board.repository.CommentRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
@@ -12,12 +14,14 @@ import org.springframework.stereotype.Service;
 public class CommentService {
 
     private final CommentRepository commentRepository;
+    private final BoardRepository boardRepository;
 
     /**
      * comment 받아서 save 하고 모댓글이면 groupId 설정 후 comment 반환
      * @param comment
      * @return 저장한 comment
      */
+    @Transactional
     public Comment saveComment(Comment comment) {
         Comment savedComment = commentRepository.save(comment);
 
@@ -26,6 +30,9 @@ public class CommentService {
             savedComment.setGroupId(comment.getCommentId());
             commentRepository.setGroupId(comment.getCommentId(), comment.getGroupId());
         }
+
+        // 댓글 수 증가
+        boardRepository.addCommentCnt(comment.getBoardId());
 
         return savedComment;
     }
@@ -42,13 +49,30 @@ public class CommentService {
     }
 
     /**
-     * comment 의 대댓글 먼저 삭제하고, 댓글 삭제 후 삭제한 열 갯수 반환
-     * @param id
-     * @return 삭제한 열 갯수
+     * comment 의 대댓글 먼저 삭제하고, 댓글 삭제 후 삭제한 댓글 갯수 반환
+     * @param id commentId
+     * @return 삭제한 댓글 갯수
      */
-    public int deleteComment(Long id) {
-        commentRepository.deleteReply(id);
-        return commentRepository.delete(id);
+    @Transactional
+    public boolean deleteComment(Long id) {
+        Comment findComment = commentRepository.findByCommentId(id);
+        int originCommentCount = 0;
+        int deletedReplyCount = 0;
+
+        if (findComment.getGroupDepth() == 0) {
+            // 모댓글 이면 답글 포함 카운트
+            originCommentCount = commentRepository.countTotalCommentWithBoardIdAndGroupId(findComment.getBoardId(), findComment.getCommentId());
+            deletedReplyCount = commentRepository.deleteReply(id); // 답글삭제
+        } else {
+            // 대댓글 이면 카운트 = 1
+            originCommentCount = 1;
+        }
+
+        int deletedCommentCount = commentRepository.delete(id); // 댓글 삭제
+        boardRepository.subtractCommentCnt(findComment.getBoardId(), deletedCommentCount + deletedReplyCount); // 댓글 수 감소
+
+//        log.info("deleteComment() groupDepth = {}, origin={}, deletedComment={}, deletedReply={}", findComment.getGroupDepth(), originCommentCount, deletedCommentCount, deletedReplyCount);
+        return originCommentCount == deletedCommentCount + deletedReplyCount;
     }
 
 }
